Added more list functionality and accessibility
Added and tested the send to web button so that when it is pressed you
are taken to a screen with all of the matches. Added to all lists the
functionality of clicking an item on the list to select that one to send
to the web. Also added the teams list where a list of teams in order of
team number is shown, and when one is clicked it should take the user to
a page with a list of the matches that that team has played in, at least
the ones that are stored on the device.