Added response to search string
Made it so that if an extra is attached to the activity under "SEARCH" the activity will check if it has a t or an m. In the case of t, the number attached to the string will be added to the string search, and the query for the listview will be a query of all data of that team. In the case of m, the number attached to the string will be attached to the search string, and the query for the listview will be a query for all of the data of that certain match number.